% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{anova.compois}
\alias{anova.compois}
\title{Teste de Razão de Verossimilhanças}
\usage{
\method{anova}{compois}(object, ...)
}
\arguments{
\item{object}{O objeto produzido por \code{\link{cmp}}}

\item{...}{Demais modelos a serem submetidos ao teste de razão de
verossimilhança}
}
\value{
Uma lista, \code{list} de componentes utilizados para exibir
    o TRV. Objeto de classe \code{anova.compois}
}
\description{
Realiza o teste de razão de verossimilhanças (TRV) para
    modelos COM-Poisson aninhados. Quando executado em um único
    modelo o TRV será feito com os modelos subjacentes ao informado,
    cujo serão removidos os efeitos preditores um a um. Caso
    executado com mais de um modelo o TRV será feito entre os modelos
    informados.
}
\author{
Eduardo E. R. Junior, \email{edujrrib@gmail.com} e Walmes
    M. Zeviani
}

