% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{bromelia}
\alias{bromelia}
\title{Alternative substrats for bromeliad production}
\format{A \code{\link[tibble]{tibble}} with 120 observations and 4
  colums:

\itemize{

\item \code{treat}: Factor indicating the alternative substrates;
\item \code{time}: Days after planting;
\item \code{block}: Factor indicating block;
\item \code{nleaves}: The median number of leaves.

}}
\usage{
data(bromelia, package = "cmpreg")
}
\description{
This dataset comes from a randomized experiment
  conducted in a greenhouse in four blocks design with objective of
  evaluate five different recipients of alternative substrates for
  bromeliads (Kanashiro, 2008). All treatments contained peat and
  perlite and differed in the third component: Pinus bark, Eucalyptus
  bark, Coxim, coconut fiber and Xaxim. The response variable was the
  number of leaves per experimental unit (pot with initially eight
  plants), which was registered at 4, 173, 229, 285, 341, and 435
  days after planting.
}
\keyword{datasets}
