% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{cholV_eta}
\alias{cholV_eta}
\title{Obtenção Pontual e Intervalar dos Preditores Lineares}
\usage{
cholV_eta(V, X, b, qn)
}
\arguments{
\item{V}{Matriz de variância e covariância das estimativas dos
parâmetros da parte de locação do modelo, necessário para
calcular a média.}

\item{X}{Matriz de funções lineares para obter \eqn{\hat{eta} = X
\hat{\beta}}, em que \eqn{\beta} são os coeficientes da porção de
locação.}

\item{b}{Estimativas dos parâmetros da parte de locação, ou seja,
\eqn{\hat{\beta}}.}

\item{qn}{Quantis da distribuição normal padrão apropriados para um
intervalo de confiança conhecida.}
}
\value{
Um vetor se \code{length(qn) == 1} e uma matriz caso
   contrário.
}
\description{
Função para obter o valores de \eqn{\eta = X\beta} que
   é o preditor da parte de locação do modelo de regressão,
   incluindo o intervalo de confiança para \eqn{\eta}, caso
   corretamente especificado pelo argumento \code{qn}.
}
\author{
Walmes Zeviani, \email{walmes@ufpr.br}.
}

