% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{cmp.predict}
\alias{cmp.predict}
\title{Valores preditos pelo Modelo Conway-Maxwell-Poisson de efeitos
    fixos}
\usage{
cmp.predict(object, newdata, type = c("link", "response"),
  interval = c("none", "confidence"), level = 0.95)
}
\arguments{
\item{object}{O objeto produzido por \code{\link{cmp}}}

\item{newdata}{Valores das covariáveis para predição, em formato de
\code{data.frame} ou \code{matrix}, respeitando a nomenclatura
das variáveis que compõem o modelo.}

\item{type}{A escala da predição. Assume valores \code{link} para a
escala logarítmica do parâmetro \eqn{\lambda} e \code{response}
para escala da contagem.}

\item{interval}{Para construção dos intervalos de confiança. Assume
valores \code{none} onde o intervalo não é calculado e
\code{confidence} para construção de intervalos de confiança para
a média da contagem sob o nível de confiança \code{level}.}

\item{level}{Nível de confiança para o intervalo predito.}
}
\value{
Um vetor, \code{vector} de valores preditos pelo modelo
    COM-Poisson, no caso de \code{interval = "none"} e uma matriz,
    \code{matrix} com os valores ajustados, inferiores e superiores
    (\emph{lwr}, \emph{fit} e \emph{upr} respectivamente) no caso de
    \code{interval = "confidence"}.
}
\description{
Calcula os valores preditos pelo modelo COM-Poisson
    ajustado na escala da função de ligação \eqn{\log(\lambda)}
    (default), ou na escala da média da contagem, que neste caso não
    é expressamente definidas, mas é função dos parâmetros
    \eqn{\lambda} e \eqn{\nu} da distribuição. Intervalos de
    confiança para a média da contagem ou para os parâmetros
    \eqn{\lambda}'s estimados também estão implementados.
}
\author{
Eduardo E. R. Junior, \email{edujrrib@gmail.com}
}

