% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{cmp}
\alias{cmp}
\title{Fitting COM-Poisson models with varying dispersion}
\usage{
cmp(formula, dformula = ~1, data, ...)
}
\arguments{
\item{formula}{An object of class "\code{\link[stats]{formula}}"
describe the model for mean.}

\item{dformula}{An object of class "\code{\link[stats]{formula}}"
describe the model for dispersion.}

\item{data}{An optional data frame containing the variables in the
model. If not found in data, the variables are taken from
\code{environment(formula)}.}

\item{...}{Arguments to be used by \code{\link{cmp_fit}}.}
}
\value{
An object of class \code{cmpreg}.
}
\description{
Fitting COM-Poisson models with varying dispersion
}
\author{
Eduardo Jr <edujrrib@gmail.com>
}
