% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{cmp_fit}
\alias{cmp_fit}
\title{Maximize the COM-Poisson log-likelihood function}
\usage{
cmp_fit(X, Z, y, start = NULL, method = c("BFGS", "Nelder-Mead", "CG",
  "L-BFGS-B", "SANN"), lower = -Inf, upper = Inf, hessian = TRUE,
  control = list())
}
\arguments{
\item{X}{Design matrix related to the (approximate) mean parameter
\eqn{\mu = \exp(X \beta)}.}

\item{Z}{Design matrix related to the dispersion parameter \eqn{\nu =
\exp(Z \gamma)}.}

\item{y}{Vector of observed count data.}

\item{start}{Initial parameters}

\item{method}{Argument passed to \code{\link[stats]{optim}(...)}.}

\item{lower}{Argument passed to \code{\link[stats]{optim}(...)}.}

\item{upper}{Argument passed to \code{\link[stats]{optim}(...)}.}

\item{hessian}{Argument passed to \code{\link[stats]{optim}(...)}.}

\item{control}{Argument passed to \code{\link[stats]{optim}(...)}.}
}
\value{
A list with estimated parameters and hessian matrix
}
\description{
Maximize the COM-Poisson log-likelihood function
}
\author{
Eduardo Jr <edujrrib@gmail.com>
}
