% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inference.R, R/methods.R, R/predict.R
\name{print.summary.cmpreg}
\alias{print.summary.cmpreg}
\alias{cmpreg-methods}
\alias{print.cmpreg}
\alias{logLik.cmpreg}
\alias{coef.cmpreg}
\alias{vcov.cmpreg}
\alias{model.matrix.cmpreg}
\alias{fitted.cmpreg}
\title{Method for 'cmpreg' objects}
\usage{
\method{print}{summary.cmpreg}(x, digits = max(3L, getOption("digits") -
  3L), ...)

\method{print}{cmpreg}(x, digits = max(3L, getOption("digits") - 3L),
  ...)

\method{logLik}{cmpreg}(object, ...)

\method{coef}{cmpreg}(object, what = c("mean", "dispersion", "all"), ...)

\method{vcov}{cmpreg}(object, ...)

\method{model.matrix}{cmpreg}(object, ...)

\method{fitted}{cmpreg}(object, what = c("mean", "dispersion", "all"),
  ...)
}
\arguments{
\item{x}{and object of class \code{cmpreg}.}

\item{digits}{minimal number of _significant_ digits, see
\code{\link[base]{print.default}}.}

\item{...}{Currently not used.}

\item{object}{an object of class \code{cmpreg}.}

\item{what}{a character indicating which parameter coefficients are
required, parameters for the \code{"mean"} or for the
\code{"dispersion"} model. If \code{"all"}, a list with
coefficients for the \code{mean} and for the \code{dispersion}
model is returned.}
}
\value{
.
}
\description{
Method for 'cmpreg' objects
}
\author{
Eduardo Jr <edujrrib@gmail.com>
}
