% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{cmptest}
\alias{cmptest}
\title{Teste de Hipóteses para o parâmetro \eqn{phi}}
\usage{
cmptest(...)
}
\arguments{
\item{...}{Uma sequência de modelos em que se realizará o teste.}
}
\description{
Realiza um teste de hipóteses bilateral para hipótese
    nula \eqn{H_0:\, \phi = 0}, ou seja, se o modelo COM-Poisson pode
    ser reduzido ao Poisson.
}
\author{
Eduardo Junior, \email{edujrrib@gmail.com}
}

