% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.compois}
\alias{coef.compois}
\title{Estimativas dos parâmetros}
\usage{
\method{coef}{compois}(object, ...)
}
\arguments{
\item{object}{O objeto produzido por \code{\link{cmp}}}

\item{...}{Argumentos adicionais para métodos S3}
}
\value{
Uma lista, \code{list} nomeada em que
\describe{
\item{betas}{É um vetor com o(s) coeficiente(s) estimados para o
    preditor linear.}
\item{nu}{É o valor estimado para o coeficiente \code{nu} do modelo
    COM-Poisson.}
}
}
\description{
Extrai os coeficientes do modelo COM-Poisson ajustado
}
\author{
Eduardo E. R. Junior, \email{edujrrib@gmail.com}
}

