% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{compute_logz}
\alias{compute_logz}
\title{Compute COM-Poisson normalizing constante in log scale}
\usage{
compute_logz(loglambda, nu)
}
\arguments{
\item{loglambda}{A vector of logarithm of the \eqn{\lambda}
parameter.}

\item{nu}{A vector of dispersion parameters \eqn{\nu}.}
}
\value{
A vector with the logarithms of the resulting
  constants.
}
\description{
Computes the normalizing constant in the log scale
  usign the LogSumExp trick to avoid numerical issues (See details).
}
\details{
\code{logspace_add(a, b) = log(exp(a) + exp(b)) = b + log(1 +
  exp(a - b))}, where \code{b > a}.
}
\references{
Wikipedia, LogSumExp. \url{http://rstudio.com}.
}
\author{
Eduardo Jr <edujrrib@gmail.com>
}
