% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{cotton}
\alias{cotton}
\title{Artificial defoliation in cotton phenology}
\format{A \code{\link[tibble]{tibble}} with 125 observations and 4
  colums:

\itemize{
\item \code{stage}: A factor with the (phenological) growth stages;
\item \code{defol}: Numerical with the defoliation levels (percent in
  leaf area removed with scissors);
\item \code{rept}: Indexes of repetition;
\item \code{bolls}: Number of bolls produced at harvest of cotton.

}}
\usage{
data(cotton, package = "cmpreg")
}
\description{
Cotton production can be drastically reduced by attack
  of defoliating insects. Depending on the growth stage, the plants
  can recover from the caused damage and keep production not affected
  or can have the production reduced by low intensity defoliation. In
  order to study the recovery of cotton plants (\emph{Gossypium
  hirsutum}) in terms of production, Silva (2012) conducted a
  greenhouse experiment under a completely randomized design with
  five replicates. The experimental unity was a pot with two plants
  and it was recorded the number of cotton bolls at five artificial
  defoliation levels (0\%, 25\%, 50\%, 75\%, and 100\%) and five
  growth stages: vegetative, flower-bud, blossom, fig and cotton
  boll.
}
\references{
Silva, A. M., Degrande, P. E., Suekane, R., Fernandes,
  M. G., Zeviani, W. M. (2012). Impacto de diferentes niveis de
  desfolha artificial nos estagios fenologicos do
  algodoeiro. \strong{Revista de Ciencias Agrarias}, 35(1), 163–172.
}
\keyword{datasets}
