% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{cottonBolls2}
\alias{cottonBolls2}
\title{Capulhos de Algodão em Função da Exposição à Mosca Branca}
\format{Um \code{data.frame} com 60 observações e 8 variáveis.
    \describe{

\item{\code{dexp}}{Inteiro com 6 valores que representa os dias de
    exposição à alta infestação de Mosca-Branca.}

\item{\code{vaso}}{Fator que indica o vaso no qual foram mensurados
    os componentes de produção do algodão.}

\item{\code{planta}}{Fator que indica a planta na qual foram
    mensurados os componentes de produção do algodão.}

\item{\code{alt}}{Altura da planta, mensurada em centímetros.}

\item{\code{pesocap}}{Peso dos capulhos de algodão, mesurado para
    cada vaso (que contém duas plantas). No \code{data.frame} somente
    a primeira planta do vaso contém a observação do peso.}

\item{\code{nerep}}{Contagem do número de estruturas reprodutivas da
    planta.}

\item{\code{ncapu}}{Contagem do número de capulhos produzidos.}

\item{\code{nnos}}{Contagem do número de nós da planta.}

}}
\description{
Experimento conduzido na Universidade Federal da Grande
    Dourados (UFGD) em 2007, cujo objetivo foi avaliar os impactos da
    exposição de plantas à alta infestação de Mosca-Branca
    \emph{Bemisia tabaci} em componentes de produção do algodão. No
    experimento, plantas de algodão foram expostas à alta infestação
    da praga por períodos diferentes e ao final do experimento
    avaliou-se o número de capulhos produzidos, o número de
    estruturas reprodutivas, o número de nós, a altura da planta e o
    peso dos capulhos por vaso. A condução do estudo deu-se via
    delineamento interamente casualizado com 5 vasos, contendo duas
    plantas, para cada período de exposição.
}
\examples{

data(cottonBolls2)

library(lattice)

# Número de estruturas reprodutivas
xyplot(nerep ~ dexp | vaso,
       data = cottonBolls2,
       jitter.x = TRUE,
       type = c("p", "g", "smooth"))

# Número de capulhos produzidos
xyplot(ncapu ~ dexp | vaso,
       data = cottonBolls2,
       jitter.x = TRUE,
       type = c("p", "g", "smooth"))

# Número de nós
xyplot(nnos ~ dexp | vaso,
       data = cottonBolls2,
       jitter.x = TRUE,
       type = c("p", "g", "smooth"))

# Altura das plantas
xyplot(alt ~ dexp | vaso,
       data = cottonBolls2,
       jitter.x = TRUE,
       type = c("p", "g", "smooth"))

# Peso dos capulhos nos vasos
da <- cottonBolls2[complete.cases(cottonBolls2), ]
xyplot(pesocap ~ dexp,
       data = cottonBolls2,
       jitter.x = TRUE,
       type = c("p", "g", "smooth"))
}
\keyword{dataset}
\keyword{efeito-aleatório}
\keyword{subdispersão}

