% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{eggs}
\alias{eggs}
\title{Produção de Ovos por Galinha}
\format{Um \code{data.frame} com 2100 observações e 7 variáveis.
    \describe{

\item{\code{periodo}}{Fator com cinco níveis que indica o período, de
    14 dias, do qual a observação provém.}

\item{\code{box}}{Fator com cinco níveis que indica o box, com seis
    gaiolas de 10 galinhas, do qual a observação provém. }

\item{\code{luz}}{Fator com cinco níveis (amarelo, azul, branco,
    verde e vermelho) que indica a cor da luz artificial sob
    incidência.}

\item{\code{gaiola}}{Fator com seis níveis que indica a gaiola,
    com 10 galinhas, da qual a observação provém.}

\item{\code{dia}}{Inteiro que indica o dia de observação no
    respectivo período.}

\item{\code{ovos}}{Contagem do número de ovos produzidos por 10
    galinhas na respectiva parcela experimental (gaiola).}

\item{\code{massa}}{Massa dos ovos produzidos na respectiva parcela
    experimental (gaiola), mensurada em gramas.}

}}
\description{
Dados provenientes de um experimento com 300 galinhas
    poedeiras da genética \emph{Isa Brown}, submetidas à regimes de
    iluminação contínua de 17 horas por dia (12h natural e 5h
    artificial) em regime 17L:7E (luz: escuridão) com diferentes
    fontes para luz artificial durante 70 dias. O objetivo do estudo
    foi avaliar o impacto de diferentes cores de luz na produção de
    ovos. Para esta avaliação o experimento foi conduzido sob o
    delineamento de quadrado latino com cinco tratamentos (cinco
    cores), divididos em cinco períodos e cinco parcelas com seis
    repetições de dez aves em cada parcela. As variáveis de interesse
    foram número de ovos produzidos e a massa dos ovos.
}
\examples{

data(eggs)

# 6 repetições de cada parcela experimental (combinação de período, luz
# e box) totalizando 84 observações em cada parcela. Experimento
# balanceado
ftable(xtabs(~periodo + luz + box, data = eggs))

library(latticeExtra)

# Contagem de ovos
useOuterStrips(
    xyplot(ovos ~ dia | luz + periodo,
           data = eggs,
           jitter.x = TRUE,
           groups = gaiola,
           type = c("p", "smooth")))

# Massa de ovos
useOuterStrips(
    xyplot(massa ~ dia | luz + periodo,
           data = eggs,
           jitter.x = TRUE,
           groups = gaiola,
           type = c("p", "smooth")))

# Relação das variáveis de interesse
xyplot(massa ~ ovos,
       data = eggs,
       type = c("p", "g", "smooth"))

# Agrupando as contagens a cada 14 dias
da <- aggregate(ovos ~ periodo + box + luz + gaiola,
                data = eggs, FUN = sum)

xyplot(ovos ~ periodo,
       groups = luz,
       data = da,
       jitter.x = TRUE,
       type = c("p", "g", "smooth"))
}
\references{
Borille, Rodrigo, Garcia, Rodrigo G., Nääs, Irenilza A.,
    Caldara, Fabiana R., & Santana, Mayara R.. (2015). Monochromatic
    light-emitting diode (LED) source in layers hens during the
    second production cycle. Revista Brasileira de Engenharia
    Agrícola e Ambiental, 19(9),
    877-881. https://dx.doi.org/10.1590/1807-1929/agriambi.v19n9p877-881

Borille, Rodrigo (2013). Led de diferentes cores como alternativa
    sustentável para iluminação de poedeiras comerciais. Dissertação
    de Mestrado, Universidade Federal da Grande Dourados, Grande
    Dourados, MS, Brasil.
    Dissertação.
}
\keyword{efeito-aleatório}

