% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_compois.R
\name{fitted.compois}
\alias{fitted.compois}
\title{Valores Ajustados pelo Modelo Conway-Maxwell-Poisson}
\usage{
\method{fitted}{compois}(object, ...)
}
\arguments{
\item{object}{O objeto produzido por \code{\link{glm_cmp}}}

\item{...}{Argumentos adicionais para métodos S3}
}
\value{
Um vetor, \code{vector} de valores de contagem ajustados pelo
    modelo COM-Poisson
}
\description{
Contagens ajustadas pelo Modelo Conway-Maxwell-Poisson
}
\author{
Eduardo E. R. Junior, \email{edujrrib@gmail.com}
}

