% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm_comp.R
\name{glm_cmp}
\alias{glm_cmp}
\title{Estimação do Modelo Conway-Maxwell-Poisson}
\usage{
glm_cmp(formula, data, ...)
}
\arguments{
\item{formula}{Um objeto da classe \code{\link{formula}}. Se
necessária a inclusão de \emph{offset} deve-se indicá-lo como
\code{\link{offset}}}

\item{data}{Um objeto de classe \code{data.frame}, cujo contém as
variáveis descritas na \code{formula}}

\item{...}{Argumentos opcionais do framework de maximização numérica
\code{\link{optim}}}
}
\value{
Uma lista com de valores retornados da \code{\link{optim}}
}
\description{
Estima os parâmetros de um modelo COM-Poisson sob a
    otimização da função de log-verossimilhança. A sintaxe
    assemelha-se com a função \code{\link{glm}} (Generalized Linear
    Models).
}
\author{
Eduardo E. R. Junior, \email{edujrrib@gmail.com}
}
\seealso{
\code{\link[tccPackage]{ll_cmp}}
}

