% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{hurdle.predict}
\alias{hurdle.predict}
\title{Valores preditos pelo Modelo Hurdle Conway-Maxwell-Poisson}
\usage{
hurdle.predict(object, newdata, type = c("response", "count", "zero", "prob"),
  at = NULL)
}
\arguments{
\item{object}{O objeto produzido por \code{\link{hurdlecmp}}}

\item{newdata}{Valores das covariáveis para predição, em formato de
\code{data.frame} ou \code{matrix}, respeitando a nomenclatura
das variáveis que compõem o modelo. Lembre-se de qque como agora
temos dois processos atuantes, a definição dessa matriz deve
contemplar os preditor para ambas as componentes.}

\item{type}{A escala da predição. Assume valores \code{response} para a
média da contagem considerando o modelo combinado, \code{prob},
para as probabilidades calculadas para \code{at} sob o modelo
ajustado, \code{zero} para retornar o valor de \eqn{\hat{\pi}}, a
probabilidade prevista pelo modelo Binomial para y>0 e
\code{count} para a média da variável COM-Poisson desconsiderando
a componente de contagens nulas.}

\item{at}{Intervalos de valores inteiros sob os quais calcular-se-á
as probabilidades descritas pelo modelo. Esse argumento só tem
efeito quando \code{type = "prob"}.}
}
\value{
Os valores preditos pelo modelo, em formato de vetor para
    \code{type = c("link", "response", "count", "zero")} e de forma
    matricial para \code{type = "prob"}.
}
\description{
Calcula os valores preditos pelo modelo Hurdle
    COM-Poisson ajustado. São disponíveis os valores preditos na
    escala da contagem ou ainda as probabilidades preditas em um
    intervalo \code{at}. Adicionalmente disponibilizamos predições
    para os componentes zero e de contagem não nula. Nesses outros
    dois tipos de predição o valor do preditor linear \eqn{\eta} é
    escalonado para a média da distribuição considera, Binomial e
    COM-Poisson para as componentes zero e de contagem
    respectivamente.
}
\author{
Eduardo E. R. Junior, \email{edujrrib@gmail.com}
}

