% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hurdlecmp.R
\name{hurdlecmp}
\alias{hurdlecmp}
\title{Ajuste de um Modelo Hurdle Conway-Maxwell-Poisson}
\usage{
hurdlecmp(formula, data, sumto = NULL, ...)
}
\arguments{
\item{formula}{Um objeto da classe \code{\link{formula}}. O preditor
linear do modelo para contagens 0 e para contagens acima de zero
deve ser separado pelo operador \code{|}, e.g. esspecificando
\code{y ~ x1 | x2} o preditor do modelo Binomial (para contagens
0) considerará apenas \code{x2} e o preditor para o modelo
COM-Poisson somente o \code{x1}. Se não for especificado o
operador \code{|} o mesmo proditor será considerado para ambas as
partes do modelo.}

\item{data}{Um objeto de classe \code{data.frame}, cujo contém as
variáveis descritas na \code{formula}}

\item{sumto}{Número de incrementos a serem considerados para a
cálculo das constantes normalizadoras. Como padrão, para cálculo
dessa constante faz-se uso de um processo iterativo, porém esse
processo é demasiadamente demorado quando se ajusta os modelos
via otimização numérica. Portanto indicar esse valor tornará o
procedimento de estimação dos parâmetros mais veloz.}

\item{...}{Argumentos opcionais do framework de maximização numérica
\code{\link[bbmle]{mle2}}}
}
\value{
Um objeto de classe \code{mle2}, retornado da função de
    \code{\link[bbmle]{mle2}}, usada para ajuste de modelos por
    máxima verossimilhança.
}
\description{
Estima os parâmetros de um modelo Hurdle COM-Poisson sob
    a otimização da função de log-verossimilhança. A sintaxe
    assemelha-se com a função \code{\link[pscl]{hurdle}} (Hurdle
    Models for Count Data Regression).
}
\author{
Eduardo E. R. Junior, \email{edujrrib@gmail.com}
}

