% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedcmp.R
\name{laplace}
\alias{laplace}
\title{Integração Numérica por meio do método de Laplace}
\usage{
laplace(funcao, n.dim, log = TRUE, ...)
}
\arguments{
\item{funcao}{Uma função cujo qual se tem interesse na obtenção do
valor de sua integral.}

\item{n.dim}{Número de dimensões a ser integral avaliada.}

\item{log}{Argumento lógico se a escala retorna é logarítmica ou
não. Por default \code{log = TRUE}.}

\item{...}{Argumentos adicionais passados a função
\code{\link{optim}}.}
}
\description{
Aproxima a integral de uma função pelo método de
    Laplace, ou seja, \deqn{\int f(x) e^{-x^2} dx =
    (2\pi)^{n/2}|{Q}|''(\hat{x})^{-1/2}e^{Q(\hat{x})}}.
}
\author{
Eduardo E. R. Junior, \email{edujrrib@gmail.com},
    implementação original em Ribeiro Jr et al. (2012).
}

