% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitting.R
\name{llcmp_fixed}
\alias{llcmp_fixed}
\alias{llcmp}
\title{Negative of the log-likelihood function from the
  reparametrized COM-Poisson model}
\usage{
llcmp_fixed(beta, gama, X, Z, y)

llcmp(params, X, Z, y)
}
\arguments{
\item{beta}{A vector of \eqn{\beta} parameter.}

\item{gama}{A vector of \eqn{\gamma} parameter.}

\item{X}{Design matrix related to the (approximate) mean parameter
\eqn{\mu = \exp(X \beta)}.}

\item{Z}{Design matrix related to the dispersion parameter \eqn{\nu =
\exp(Z \gamma)}.}

\item{y}{Vector of observed count data.}

\item{params}{A vector of the model parameters \code{params} =
\code{c(beta, gama)}.}
}
\value{
The computed log-likelihood function.
}
\description{
Compute the negative of the logarithm of the likelihood
  function for a set of observations \code{y} from the reparametrized
  COM-Poisson model given the \eqn{\mu} and \eqn{\nu} parameters (see
  Details).
}
\details{
The log-likelihood function is given by
  \deqn{\ell(\beta,\nu) = \sum y \log\lambda - \nu\log y! -
  \log[Z(\lambda,\nu)],} where \deqn{\log\lambda = \nu \log\left(\mu
  - \frac{\nu-1}{2\nu}\right),} and \eqn{\log[Z(\lambda, \nu)]} is a
  normalizing constant computed in log space to avoid numerical
  issues (see \code{\link{compute_logz}}).
}
\author{
Eduardo Jr <edujrrib@gmail.com>
}
