% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmp.R
\name{llcmp}
\alias{llcmp}
\title{Log-Verossimilhança do Modelo Conway-Maxwell-Poisson}
\usage{
llcmp(betas, phi, y, X, offset = NULL)
}
\arguments{
\item{betas}{Um vetor de estimativas para os parâmetros de regressão
da distribuição Conway-Maxwell-Poisson.}

\item{phi}{Um valor estimado para o parâmetro de dispersão considerado
na distrbuição Conway-Maxwell-Poisson \eqn{\phi = \log{\nu}}}

\item{y}{Um vetor de contagens, considerado como variável resposta}

\item{X}{A matriz de delineamento do modelo}

\item{offset}{Um vetor de valores a serem adicionados ao preditor
linear}
}
\value{
O valor da log-verossimilhança do modelo
    Conway-Maxwell-Poisson com os parâmetros e dados informados
}
\description{
Calcula a log-verossimilhança de um modelo COM-Poisson
    considerando os dados e as estimativas dos parâmetros informadas.
}
\details{
A função de log-verossimilhança toma a forma: \deqn{-Z - y *
    \lambda - \nu \log{y!}}, onde \eqn{Z = \sum
    \frac{\lambda^i}{(i!)^\nu}}
}
\author{
Eduardo E. R. Junior, \email{edujrrib@gmail.com}
}
\seealso{
\code{\link[tccPackage]{cmp}}
}

