% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hurdlecmp.R
\name{llhurdle}
\alias{llhurdle}
\title{Log-Verossimilhança do Modelo Hurdle Conway-Maxwell-Poisson}
\usage{
llhurdle(params, y, Xc, Xz, sumto = NULL)
}
\arguments{
\item{params}{Um vetor de estimativas para os parâmetros da
distribuição Conway-Maxwell-Poisson truncada em zero. O verto
deve ser nomeado com prefixo \code{count}, para os parâmetros do
modelo COM-Poisson e \code{zero} para os que se referem ao modelo
Binomial associado a contagens 0.}

\item{y}{Um vetor de contagens, considerado como variável resposta}

\item{Xc}{A matriz de delineamento do modelo para contagens \eqn{y >
0}}

\item{Xz}{A matriz de delineamento do modelo para contagens nulas
\eqn{y = 0}}

\item{sumto}{Número de incrementos a serem considerados para a
cálculo das constantes normalizadoras. Como padrão, para cálculo
dessa constante faz-se uso de um processo iterativo, porém esse
processo é demasiadamente demorado quando se ajusta os modelos
via otimização numérica. Portanto indicar esse valor tornará o
procedimento de estimação dos parâmetros mais veloz.}
}
\value{
O valor da log-verossimilhança do modelo
    Hurdle Conway-Maxwell-Poisson com os parâmetros e dados informados
}
\description{
Calcula a log-verossimilhança de um modelo Hurdle
    COM-Poisson considerando os dados e as estimativas dos parâmetros
    informadas. As distribuições consideradas para a porção de
    valores nulos (\eqn{y = 0}) e para a contagem truncada \eqn{y >
    0} são Binomial e COM-Poisson respectivamente.
}
\details{
A função de log-verossimilhança do modelo toma a forma:
    \deqn{llcmp_{hurdle} = \log(\pi_i) I_{y = 0} + ( \log(1 - \pi_i)+
    y \log(\lambda) - e^{\phi} \log(y!) - \log(Z) - \log(1 - 1/Z))
    I_{y > 0}}, onde \eqn{Z = \sum \frac{\lambda^i}{(i!)^\nu}}
}
\author{
Eduardo E. R. Junior, \email{edujrrib@gmail.com}
}
\seealso{
\code{\link[tccPackage]{hurdlecmp}}
}

