% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedcmp.R
\name{llicmp}
\alias{llicmp}
\title{Log-verossimilhança por grupo do modelo COM-Poisson misto}
\usage{
llicmp(b, params, y, X, Z, sumto = NULL)
}
\arguments{
\item{b}{Efeito aleatório nosiderada para avaliação da
log-verossimilhança.}

\item{params}{Vetor de parâmetros do modelo. Deve seguir a ordenação
    dada abaixo: \itemize{

\item{\code{phi}}{Parâmetro extra da distribuição COM-Poisson;}

\item{\code{lsigma}}{Parâmetros relacionados a matriz Z dos efeitos;
    aleatórios. Representam as \eqn{\log(\sqrt(\textrm{variâncias}))}
    na diagonal da matriz de variâncias e covariâncias \eqn{Sigma} da
    Normal q-variada considerada para os efeitos aleatórios;}

\item{\code{betas}}{Parâmetros relacionados aos efeitos fixos do
    modelo.}

}}

\item{y}{Um vetor de contagens do grupo considerado.}

\item{X}{A matriz de delineamento dos efeitos fixos do modelo no
grupo considerado.}

\item{Z}{A matriz de delineamento dos efeitos aleatórios do modelo no
grupo considerado.}

\item{sumto}{Número de incrementos a serem considerados para a
cálculo da constante normalizadora Z.}
}
\description{
Avalia a log-verossimilhança de cada grupo formado pelas
    categóricas da variável considerada como aleatória.
}

