% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedcmp.R
\name{llmixedcmp}
\alias{llmixedcmp}
\title{Log-verossimilhança Marginal do Modelo COM-Poisson Misto}
\usage{
llmixedcmp(params, form, form.X, form.Z, dados.id, sumto = NULL)
}
\arguments{
\item{params}{Vetor de parâmetros do modelo. Deve seguir a ordenação
    dada abaixo: \itemize{

\item{\code{phi}}{Parâmetro extra da distribuição COM-Poisson;}

\item{\code{lsigma}}{Parâmetros relacionados a matriz Z dos efeitos;
    aleatórios. Representam as \eqn{\log(\sqrt(\textrm{variâncias}))}
    na diagonal da matriz de variâncias e covariâncias \eqn{Sigma} da
    Normal q-variada considerada para os efeitos aleatórios;}

\item{\code{betas}}{Parâmetros relacionados aos efeitos fixos do
    modelo.}

}}

\item{form}{Um objeto da classe \code{\link{formula}}. Essa fórmula
substitue os "pipes", da fármula genérica para modelos mistos,
pelo símbolo de adição, veja \code{\link[lme4]{subbars}}.}

\item{form.X}{Um objeto da classe \code{\link{formula}}. Essa fórmula
deve representar somente os efeitos fixos do modelo, veja
\code{\link[lme4]{nobars}}.}

\item{form.Z}{Um objeto da classe \code{\link{formula}}. Essa fórmula
deve representar somente os efeitos aleatórios do modelo, veja
\code{\link[lme4]{findbars}}.}

\item{dados.id}{Um objeto de classe \code{\link{list}}. Esa lista
deve ser uma divisão dos dados, onde cada slot contém a
participação comum a um grupo identificado por níveis
considerados como efeitos aleatórios.}

\item{sumto}{Número de incrementos a serem considerados para a
cálculo da constante normalizadora Z.}
}
\description{
Avalia a log-verossimilhaça marginal de um modelo
    COM-Poisson misto. Aqui as log-verossimilhanças de cada grupo são
    integradas numericamente em relação aos efeitos aleatórios.
}

