% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_compois.R
\name{logLik.compois}
\alias{logLik.compois}
\title{Log-Verossimilhança para modelos Conway-Maxwell-Poisson}
\usage{
\method{logLik}{compois}(object, ...)
}
\arguments{
\item{object}{O objeto produzido por \code{\link{glm_cmp}}}

\item{...}{Argumentos adicionais para métodos S3}
}
\value{
Um objeto de classe \code{logLik}
}
\description{
Exibe a log-verossimilhança calculada para o modelo
    Conway-Maxwell-Poisson ajustado em objetos de classe
    \code{compois}.
}
\author{
Eduardo E. R. Junior, \email{edujrrib@gmail.com}
}

