% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{mixedcmp.predict}
\alias{mixedcmp.predict}
\title{Valores preditos pelo Modelo Conway-Maxwell-Poisson de efeitos
    mistos}
\usage{
mixedcmp.predict(object, newdata = NULL, type = c("link", "response"))
}
\arguments{
\item{object}{O objeto produzido por \code{\link{mixedcmp}}}

\item{newdata}{Valores das covariáveis para predição, em formato de
\code{data.frame}.}

\item{type}{A escala da predição. Assume valores \code{link} para a
escala logarítmica do parâmetro \eqn{\lambda} e \code{response}
para escala da contagem.}
}
\value{
Um vetor, \code{vector} de valores preditos pelo modelo
    COM-Poisson de efeitos mistos.
}
\description{
Calcula os valores preditos pelo modelo COM-Poisson
    ajustado na escala da função de ligação \eqn{\log(\lambda)}
    (default), ou na escala da média da contagem, que neste caso não
    é expressamente definida, mas é função dos parâmetros
    \eqn{\lambda} e \eqn{\nu} da distribuição.
}
\author{
Eduardo E. R. Junior, \email{edujrrib@gmail.com}
}

