% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{mixedcmp.ranef}
\alias{mixedcmp.ranef}
\title{Obtenção dos Efeitos Aleatórios no Modelo COM-Poisson Misto}
\usage{
mixedcmp.ranef(object)
}
\arguments{
\item{object}{Uma sequência de modelos em que se realizará o teste.}
}
\description{
Encontra os efeitos aleatórios que maximizam a
    log-verossimilhança do modelo  COM-Poisson de efeitos mistos.
}
\author{
Eduardo Junior, \email{edujrrib@gmail.com}
}

