% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixedcmp.R
\name{mixedcmp}
\alias{mixedcmp}
\title{Ajuste de um Modelo Misto Conway-Maxwell-Poisson}
\usage{
mixedcmp(formula, data, sumto = NULL, ...)
}
\arguments{
\item{formula}{Um objeto da classe \code{\link{formula}}. A
especificação dos efeitos aleatórios é feita neste argumento,
e.g. \code{y ~ x + (x | id)} especifica que o modelo será um
COM-Poisson de intercepto e inclinação aleatórios. Os efeitos
considerados são independentes e normalmente distribuídos.}

\item{data}{Um objeto de classe \code{data.frame}, cujo contém as
variáveis descritas na \code{formula}}

\item{sumto}{Número de incrementos a serem considerados para a
cálculo da constante normalizadora Z.}

\item{...}{Argumentos opcionais do framework de maximização numérica
\code{\link[bbmle]{mle2}}.}
}
\value{
Um objeto de classe \code{mle2}, retornado da função de
    \code{\link[bbmle]{mle2}}, usada para ajuste de modelos por
    máxima verossimilhança.
}
\description{
Estima os parâmetros de um modelo misto COM-Poisson sob
    a otimização da função de log-verossimilhança. A sintaxe
    assemelha-se com a função \code{\link[lme4]{lmer}}, ou
    \code{\link[lme4]{glmer}} (Generalized Linear Mixed-Effects
    Models).
}
\author{
Eduardo E. R. Junior, \email{edujrrib@gmail.com}
}

