% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{nitrofen}
\alias{nitrofen}
\title{Toxicity of nitrofen in aquatic systems}
\format{A \code{\link[tibble]{tibble}} with 50 observations and 2
  colums:

\itemize{

\item \code{dose}: Numeric value of the nitrofen concentration level
  (in \eqn{\mu}g/ 100 litre);
\item \code{noffs}: Number of live offspring.

}}
\usage{
data(nitrofen, package = "cmpreg")
}
\description{
Nitrofen is no longer in commercial use in the United
  States, having been the first pesticide to be withdrawn due to
  tetragenic effects (Bailer, 1994). This data set comes from an
  experiment to measure the reproductive toxicity of the herbicide,
  nitrofen, on a species of zooplankton (\emph{Ceriodaphnia
  dubia}). Fifty animals were randomized into batches of ten and each
  batch was put in a solution with a measured concentration of
  nitrofen (0, .8, 1.6, 2.35 and 3.10
  \eqn{\mu}g/100litre. Subsequently, the number of live offspring was
  recorded.
}
\references{
Bailer, A. and J. Oris (1994). Assessing toxicity of
  pollutant in aquatic systems. \strong{In Case Studies in Biometry},
  25-40.
}
\keyword{datasets}
