% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict.cmpreg}
\alias{predict.cmpreg}
\title{Predict method for COM-Poisson models}
\usage{
\method{predict}{cmpreg}(object, newdata, newmatrices = NULL,
  what = c("mean", "dispersion", "all"), type = c("link", "response"),
  se.fit = FALSE, augment_data = FALSE, ...)
}
\arguments{
\item{object}{a fitted object of class from \code{"glm"}.}

\item{newdata}{optionally, a data frame in which to look for
variables with which to predict. If omitted, the fitted linear
predictors are used.}

\item{newmatrices}{optionally, a list with named design matrices
(\code{"X"} for mean model and \code{"Z"} for dispersion model)
used to predict. If omitted, the fitted linear predictors are used.}

\item{what}{a character indicating which parameter coefficient is
required, parameters for the \code{"mean"} or for the
\code{"dispersion"} model.}

\item{type}{the type of prediction required. The default
\code{"link"} is on the scale of the linear predictors; the
alternative \code{"response"} is on the scale of the response
variable.}

\item{se.fit}{logical switch indicating if standard errors are
required. Default is \code{FALSE}.}

\item{augment_data}{logical indicating if \code{newdata} should be
augmented with the predict values and, possibly, standard
errors. Default is \code{FALSE}.}

\item{...}{currently not used.}
}
\value{
a tibble with fitted values (code{fit}) and, possibly,
  standard erros (\code{ste}). If \code{augment_data}, the result
  will be the \code{newdata} with new columns \code{fit}, fitted
  values, and code{ste}, standard errors.
}
\description{
Predict method for COM-Poisson models
}
\author{
Eduardo Jr <edujrrib@gmail.com>
}
