% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_compois.R
\name{predict.compois}
\alias{predict.compois}
\title{Valores preditos pelo Modelo Conway-Maxwell-Poisson}
\usage{
\method{predict}{compois}(object, newdata, type = c("link", "response"),
  interval = c("none", "confidence"), level = 0.95, ...)
}
\arguments{
\item{object}{O objeto produzido por \code{\link{cmp}}}

\item{newdata}{Valores das covariáveis para predição}

\item{type}{A escala da predição. Assume valores \code{link} para a
escala do parâmetro \eqn{\lambda} e \code{response} para escala
da contagem}

\item{interval}{Para construção do intervalo de confiança. Assume
valores \code{none} onde o intervalo não é calculado e
\code{confidence} para construção do intervalo de confiança para
a média da contagem sob o nível de confiança \code{level}.}

\item{level}{Nível de confiança para o intervalo predito}

\item{...}{Argumentos adicionais para métodos S3}
}
\value{
Um vetor, \code{vector} de valores preditos pelo modelo
    COM-Poisson, no caso de \code{interval = "none"} e uma matriz,
    \code{matrix} com os valores ajustados, inferiores e superiores
    (\emph{fit}, \emph{lwr} e \emph{upr} respectivamente) no caso de
    \code{interval = "confidence"}
}
\description{
Calcula os valores preditos pelo modelo COM-Poisson
    ajustado na escala da função de ligação, que neste caso não é
    expressamente definidas, mas é função dos parâmetros \eqn{\lambda}
    e \eqn{\nu} da distribuição, ou na escala da média da
    contagem. Intervalos de confiança para a média da contagem ou
    para os parâmetros \eqn{\lambda}'s estimados também estão
    implementados.
}
\author{
Eduardo E. R. Junior, \email{edujrrib@gmail.com} e Walmes
    M. Zeviani
}

