% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.anova.compois}
\alias{print.anova.compois}
\title{Exibição da ANOVA do Modelo Conway-Maxwell-Poisson}
\usage{
\method{print}{anova.compois}(x, ...)
}
\arguments{
\item{x}{O objeto produzido por \code{\link{anova.compois}}}

\item{...}{Argumentos adicionais para métodos S3}
}
\description{
Exibe a análise de log-verossimilhança entre modelos
    COM-Poisson
}
\author{
Eduardo E. R. Junior, \email{edujrrib@gmail.com}
}

