% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_compois.R
\name{print.compois}
\alias{print.compois}
\title{Exibição do Ajuste do Modelo Conway-Maxwell-Poisson}
\usage{
\method{print}{compois}(x, ...)
}
\arguments{
\item{x}{O objeto produzido por \code{\link{cmp}}}

\item{...}{Argumentos adicionais para métodos S3}
}
\description{
Exibe o ajuste do modelo Conway-Maxwell-Poisson para
    objetos de classe \code{compois}
}
\author{
Eduardo E. R. Junior, \email{edujrrib@gmail.com}
}

