% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.summary.compois}
\alias{print.summary.compois}
\title{Exibição do Resumo do Modelo Conway-Maxwell-Poisson Ajustado}
\usage{
\method{print}{summary.compois}(x, ...)
}
\arguments{
\item{x}{O objeto produzido por \code{\link{summary.compois}}}

\item{...}{Argumentos adicionais para métodos S3}
}
\description{
Exibe as estatísticas resumo do modelo ajustado
}
\author{
Eduardo E. R. Junior, \email{edujrrib@gmail.com}
}

