% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{residuals.compois}
\alias{residuals.compois}
\title{Resíduos do Modelo Conway-Maxwell-Poisson Ajustado}
\usage{
\method{residuals}{compois}(object, type = c("raw", "deviance", "pearson"),
  ...)
}
\arguments{
\item{object}{O objeto produzido por \code{\link{cmp}}}

\item{type}{Tipo de resíduo a ser calculado}

\item{...}{Argumentos adicionais para métodos S3}
}
\value{
Um vetor, \code{vector} de resíduos calculados sob o modelo
    COM-Poisson ajustado
}
\description{
Calcula os resíduos do modelo ajustado
}
\author{
Eduardo E. R. Junior, \email{edujrrib@gmail.com}
}

