% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{sitophilus}
\alias{sitophilus}
\title{Annona mucosa for control of Sitophilus zeamaus}
\format{A \code{\link[tibble]{tibble}} with 40 observations and 2
  colums:

\itemize{

\item \code{extract}: Factor indicating the extrated used in the
  solution;
\item \code{ninsect}: Number of emerged insects.

}}
\usage{
data(sitophilus, package = "cmpreg")
}
\description{
New control methods are necessary for stored grain pest
  management programs due to both the widespread problems of
  insecticide-resistance populations and the increasing concerns of
  consumers regarding pesticide residues in food products. Ribeiro
  (2013) carried out an experiment to assess the bioactivity of
  extracts of \emph{Annona mucosa} (Annonaceae) for control
  \emph{Sitophilus zeamaus} (Coleoptera: Curculionidae), a major pest
  of stored maize in Brazil. Petri dishes containing 10g of corn were
  treated with extracts prepared with different parts of \emph{Annona
  mucosa} (seeds, leaves and branches) or just water (control) were
  completely randomized with 10 replicates. Then 20 animals adults
  were placed in each Petri dish and the numbers of emerged insects
  (progeny) after 60 days were recorded.
}
\references{
Ribeiro, L. P., J. D. Vendramim, K. U. Bicalho,
  M. S. Andrade, J. B. Fernandes, R. A. Moral, and C. G. B. Demetrio
  (2013). Annona mucosa Jacq. (Annonaceae): A promising source of
  bioactive compounds against Sitophilus zeamais Mots. (Coleoptera:
  Curculionidae). \strong{Journal of Stored Products Research}, 55,
  6-14.
}
\keyword{datasets}
