% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\docType{data}
\name{soybean}
\alias{soybean}
\title{Soil moisture and potassium doses on soybean culture}
\format{A \code{\link[tibble]{tibble}} with 74 observations and 6
  colums:

\itemize{

\item \code{K}: Integer value indicated the potassium fertilization
  dose (in mg dm\eqn{^{-3}});
\item \code{water}: Numerical value of amount of water in the soil
  (soil moisture in percent);
\item \code{block}: Fatcor indicating block;
\item \code{seeds}: Number of bean seeds at pot;
\item \code{vpods}: Number of viable pods at pot;
\item \code{tpods}: Total of pods at pot.

}}
\usage{
data(soybean, package = "cmpreg")
}
\description{
A study of potassium doses and soil moisture levels on
  soybean (\emph{Glicine Max}) production. The tropical soils are
  usually poor in potassium (K) and demand potassium fertilization
  when cultivated with soybean to obtain satisfactory yields. Soybean
  production is affected by long exposition to water deficit. As
  potassium is a nutrient involved in the water balance in plant, by
  hyphotesis, a good supply of potassium avoids to reduce
  production. To evaluate the effects of potassium doses and soil
  humidity levels on soybean production, Serafim (2012) conducted a
  \eqn{5\times 3} factorial experiment in a randomized complete
  block design with 5 replicates. Five different potassium doses (0,
  30, 60, 120 and 180 \eqn{\times} mg dm\eqn{^{-3}}) were applied to
  the soil and soil moisture levels were controlled at (37.5, 50, and
  62.5\%). The experiment was carried out in a greenhouse and the
  experimental units were pots with two plants in each.
}
\references{
Serafim, M. E., F. B. Ono, W. M. Zeviani, J. O. Novelino,
  and J. V. Silva (2012). Umidade do solo e doses de potassio na
  cultura da soja. \strong{Revista Ciencia Agronomica}, 43(2),
  222-227.
}
\keyword{datasets}
