% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.compois}
\alias{summary.compois}
\title{Resumo do Modelo Conway-Maxwell-Poisson Ajustado}
\usage{
\method{summary}{compois}(object, ...)
}
\arguments{
\item{object}{O objeto produzido por \code{\link{cmp}}}

\item{...}{Argumentos adicionais para métodos S3}
}
\value{
Uma lista, \code{list} de componentes utilizados para resumir
    o ajuste do modelo. Objeto de classe \code{summary.compois}
}
\description{
Calcula as estatísticas resumo do modelo ajustado
}
\author{
Eduardo E. R. Junior, \email{edujrrib@gmail.com} e Walmes
    M. Zeviani
}

