% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods_compois.R
\name{vcov.compois}
\alias{vcov.compois}
\title{Matriz de Variância e Covariância do Modelo
    Conway-Maxwell-Poisson Ajustado}
\usage{
\method{vcov}{compois}(object, ...)
}
\arguments{
\item{object}{O objeto produzido por \code{\link{glm_cmp}}}

\item{...}{Argumentos adicionais para métodos S3}
}
\value{
Uma matriz nomeada, \code{matrix} com as variâncias e
    covariâncias entre os parâmetros estimados
}
\description{
Calcula a matriz de variância e covariância numérica de
    um modelo COM-Poisson
}
\author{
Eduardo E. R. Junior, \email{edujrrib@gmail.com}
}

