% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmp.R
\name{cmp}
\alias{cmp}
\title{Estimação do Modelo Conway-Maxwell-Poisson}
\usage{
cmp(formula, data, sumto = NULL, ...)
}
\arguments{
\item{formula}{Um objeto da classe \code{\link{formula}}. Se
necessária a inclusão de \emph{offset} deve-se indicá-lo como
\code{\link{offset}}}

\item{data}{Um objeto de classe \code{data.frame}, cujo contém as
variáveis descritas na \code{formula}}

\item{sumto}{Número de incrementos a serem considerados para a
cálculo da constante normalizadora Z.}

\item{...}{Argumentos opcionais do framework de maximização numérica
\code{\link{optim}}}
}
\value{
Uma lista de componentes do ajuste. Objeto de classe
    \code{compois} cujo funções métodos foram implementadas.
}
\description{
Estima os parâmetros de um modelo COM-Poisson sob a
    otimização da função de log-verossimilhança. A sintaxe
    assemelha-se com a função \code{\link{glm}} (Generalized Linear
    Models).
}
\author{
Eduardo E. R. Junior, \email{edujrrib@gmail.com}
}

