% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cmp.R
\name{computez}
\alias{computez}
\title{Avaliação da Constante Normalizadora}
\usage{
computez(loglambda, phi, tol = 0.01, maxit = 500, incremento = 10)
}
\arguments{
\item{loglambda}{Valor de \eqn{\log(\lambda)}, sendo \eqn{\lambda} o
parâmetro da função de distribuição de probabilidades COM-Poisson}

\item{phi}{Parâmetro \eqn{\phi = \log{\nu}} da função de distribuição
de probabilidades COM-Poisson}

\item{tol}{Critério de parada do algoritmo, representa o valor
tolerado para a diferença do valor de \eqn{Z(\lambda, \phi)}
entre duas iterações. O valor padrão é 1e-3}

\item{maxit}{Número máximo de iterações a serem realizadas pelo
algoritmo. Se este número for atingido e o critério de tolerância
não for atendido, uma mensagem de aviso será exibida}

\item{incremento}{Número de incrementos da soma a serem considerados
a cada iteração. Padrão definido como 10, ou seja, a cada
iteração 10 incrementos são calculados.}
}
\value{
O valor da constante de normalização, \eqn{Z(\lambda, \phi)}
    da distribuição COM-Poisson
}
\description{
Calcula o valor da constante de normalização do modelo
    COM-Poisson definida por: \deqn{Z = \sum
    \frac{\lambda^i}{(i!)^\nu}}. Para melhoria dos métodos de
    estimação o parâmetro \eqn{\nu} é tomado como \eqn{\exp{\phi}} e
    todas as inferências são tomadas a partir de \eqn{\phi}
}
\author{
Eduardo E. R. Junior, \email{edujrrib@gmail.com}
}

