% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hurdlecmp.R
\name{hurdlecmp}
\alias{hurdlecmp}
\title{Estimação do Modelo Hurdle Conway-Maxwell-Poisson}
\usage{
hurdlecmp(formula, data, ...)
}
\arguments{
\item{formula}{Um objeto da classe \code{\link{formula}}. Se
necessária a inclusão de \emph{offset} deve-se indicá-lo como
\code{\link{offset}}. O preditor linear do modelo para contagens
0 e para contagens acima de zero deve ser separado pelo operador
\code{|}, por exemplo se for descrito \code{y ~ x1 | x2} o
preditor do modelo Binomial (para contagens 0) considerará apenas
\code{x2} e o preditor para o modelo COM-Poisson somente o
\code{x1}. Se não for especificado o operador \code{|} o mesmo
proditor será considerado para ambas as porções do modelo.}

\item{data}{Um objeto de classe \code{data.frame}, cujo contém as
variáveis descritas na \code{formula}}

\item{...}{Argumentos opcionais do framework de maximização numérica
\code{\link{optim}}}
}
\value{
Uma lista de componentes do ajuste. Objeto de classe
    \code{hurdlecmp} cujo funções métodos foram implementadas.
}
\description{
Estima os parâmetros de um modelo Hurdle COM-Poisson sob
    a otimização da função de log-verossimilhança. A sintaxe
    assemelha-se com a função \code{\link[pscl]{hurdle}} (Hurdle
    Models for Count Data Regression).
}
\author{
Eduardo E. R. Junior, \email{edujrrib@gmail.com}
}

