% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.R
\name{soyaBeans}
\alias{soyaBeans}
\title{Avaliação da Cultura de Soja}
\format{Um \code{data.frame} com 75 observações e 10 variáveis.
    \describe{

\item{\code{potassio}}{Número inteiro que representa a adubação
    potássica experimentada em cada parcela. Foram 5 nivéis de
    adubação considerados.}

\item{\code{agua}}{Numérico que representa o nível de água do solo
    que as parcelas foram cultivadas. Foram 3 níveis experimentados
    pouca água, água em quantidade ideal, água em abundância.}

\item{\code{bloco}}{Fator com 5 níveis que representam os blocos
    utilizados para controle de variação local.}

\item{\code{rend}}{Rendimento de grãos.}

\item{\code{peso}}{Peso de grãos.}

\item{\code{kgrao}}{Conteúdo de potássio no grão.}

\item{\code{pgrao}}{Conteúdo de fósforo no grão.}

\item{\code{ts}}{Total de sementes por planta.}

\item{\code{nvi}}{Número de vagens inviáveis.}

\item{\code{nv}}{Número de vagens total.}

}}
\source{
???
}
\description{
Experimento conduzido na Universidade Federal da Grande
    Dourados (UFGD) em casa de vegetação com a cultura de soja. Foram
    experimentados diferentes fatores de água do solo e de adubação
    potássica, em cada parcela que continha 2 plantas. Para controle
    de variação local, as parcelas foram arranjadas em blocos.
}
\details{
As variáveis \code{ts} e \code{nv} possuem alta correlação
    (0.978), pois em média são esperados 3 sementes por cada
    vagem. Então \code{ts} pode ser tratada como função de \code{nv}.
}
\keyword{(efeito}
\keyword{aleatório)}
\keyword{equidispersão}

